delete button works

+++ src/components/DeleteTask.jsx
@@ -3,17 +3,19 @@ import axios from 'axios'
 import { AppContext } from "../context/AppContext";
 
 
-const CompleteButton = props => {
+const DeleteTask = props => {
   const { setRefetch } = useContext(AppContext)
 
   const handleDelete = async () =>{
     const token = localStorage.getItem("token")
-    await axios({
-      method: 'DELETE',
-      url: `http://localhost:8080/tasks/${props.id}`,
-      headers: {Authorization: `Bearer ${token}`} 
-    })
-    setRefetch(true)
+     if (window.confirm("Delete this task?")){
+      await axios({
+        method: 'DELETE',
+        url: `http://localhost:8080/tasks/${props.id}`,
+        headers: {Authorization: `Bearer ${token}`} 
+      })
+      setRefetch(true)
+     }  
   }
 
 return(
@@ -23,4 +25,4 @@ return(
 )
 }
 
-export default CompleteButton
+export default DeleteTask
